app/producer/requirements.txt
fastapi
pydantic
kafka-python
uvicorn